Move setupBase out
