+++ adapter/build.gradle.kts
@@ -1,12 +1,8 @@
-import com.android.build.gradle.LibraryExtension
-
 plugins {
   id(libs.plugins.android.library.get().pluginId)
   id(libs.plugins.kotlin.android.get().pluginId)
   id(libs.plugins.kotlin.kapt.get().pluginId)
 }
-
-setupBase<LibraryExtension>(LibModule.Adapter)
 
 dependencies {
   api(libs.androidX.collection)
