+++ biz-detail/build.gradle.kts
@@ -4,12 +4,14 @@ plugins {
   alias(libs.plugins.moshiX)
 }
 
-setupLib(LibModule.Detail) {
+android {
   buildFeatures.compose = true
   composeOptions.kotlinCompilerExtensionVersion = libs.versions.androidX.compose.get()
 }
 
 dependencies {
+  implementation(projects.common)
+
   implementation(libs.androidX.activity.compose)
   implementation(libs.bundles.androidX.compose)
   implementation(libs.bundles.accompanist)
